pyproject.toml
[project]
name = "chat-bots-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "google-cloud-api-keys>=0.5.16",
    "google-cloud-dialogflow>=2.41.1",
    "python-dotenv>=1.1.0",
    "python-telegram-bot==13.15",
    "vk-api>=11.9.9",
]
